+++ packages/contracts-core/contracts/manager/BondingManager.sol
@@ -66,7 +66,7 @@ contract BondingManager is AgentManager, InterfaceBondingManager {
     function initialize(address origin_, address destination_, address inbox_, address summit_) external initializer {
         __AgentManager_init(origin_, destination_, inbox_);
         summit = summit_;
-        __Ownable_init();
+        __Ownable2Step_init();
         // Insert a zero address to make indexes for Agents start from 1.
         // Zeroed index is supposed to be used as a sentinel value meaning "no agent".
         _agents.push(address(0));